+++ Sandswept/Survivors/Ranger/States/Secondary/Char.cs
@@ -4,11 +4,14 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
 {
     public class Char : BaseState
     {
-        public static float damageCoefficient = 6f;
+        public static float damageCoefficient = 15f;
         public static float baseDuration = 0.5f;
+        public static float heatReduction = 0.5f;
         private float duration;
+        public float finalDamageCoefficient;
 
         private GameObject charProjectile;
+        public RangerHeatController rangerHeatController;
 
         public override void OnEnter()
         {
@@ -16,7 +19,11 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
 
             var modelTransform = GetModelTransform();
 
-            GetComponent<RangerHeatController>().currentHeat -= 50f;
+            rangerHeatController = GetComponent<RangerHeatController>();
+            float heatRemoved = rangerHeatController.currentHeat * heatReduction;
+            rangerHeatController.currentHeat -= Mathf.Max(0, heatRemoved);
+
+            finalDamageCoefficient = damageCoefficient + (heatRemoved / 2.5f);
 
             if (modelTransform)
             {
@@ -81,7 +88,7 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
             {
                 var fpi = new FireProjectileInfo()
                 {
-                    damage = damageStat * damageCoefficient,
+                    damage = damageStat * finalDamageCoefficient,
                     crit = RollCrit(),
                     damageColorIndex = DamageColorIndex.Default,
                     owner = gameObject,