+++ Sandswept/Survivors/Ranger/States/Secondary/Exhaust.cs
@@ -5,15 +5,19 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
 {
     public class Exhaust : BaseState
     {
-        public static float damageCoefficient = 2f;
+        public static float damageCoefficient = 3f;
         public static float procCoefficient = 1f;
         public static float baseDurationPerVolley = 0.3f;
         public static int baseVolleyCount = 1;
         public static float heatPerExtraVolley = 20f;
         public static float heatReduction = 0.5f;
+        public static float baseShotDistance = 200f;
+        public static float shotDistanceScaling = -15f;
+        public static float minimumShotDistance = 50f;
         public int finalVolleyCount;
         public float durationPerVolley;
         public float finalDuration;
+        public float finalShotDistance;
         public GameObject tracerEffect;
         public GameObject impactEffect;
         public RangerHeatController rangerHeatController;
@@ -26,13 +30,15 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
 
             rangerHeatController = GetComponent<RangerHeatController>();
 
+            finalShotDistance = Mathf.Max(minimumShotDistance, baseShotDistance + shotDistanceScaling * rangerHeatController.currentHeat / 5);
+
             float heatRemoved = rangerHeatController.currentHeat * heatReduction;
             rangerHeatController.currentHeat -= Mathf.Max(0, heatRemoved);
 
             durationPerVolley = baseDurationPerVolley / attackSpeedStat;
 
             finalVolleyCount = baseVolleyCount + Mathf.FloorToInt(heatRemoved / heatPerExtraVolley);
-            Main.ModLogger.LogError("final volley count is " + finalVolleyCount);
+            // Main.ModLogger.LogError("final volley count is " + finalVolleyCount);
 
             Util.PlaySound("Play_mage_m2_charge", gameObject);
             Util.PlaySound("Play_mage_m2_charge", gameObject);
@@ -96,11 +102,11 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
 
         public IEnumerator FireShot()
         {
-            Main.ModLogger.LogError("running FireShot()");
+            // Main.ModLogger.LogError("running FireShot()");
             for (int i = 0; i < finalVolleyCount; i++)
             {
                 yield return new WaitForSeconds(durationPerVolley / (i + 1));
-                Main.ModLogger.LogError("shooting volley");
+                // Main.ModLogger.LogError("shooting volley");
                 characterBody.SetAimTimer(0.3f);
 
                 var aimDirection = GetAimRay().direction;
@@ -121,7 +127,8 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
                     owner = gameObject,
                     isCrit = RollCrit(),
                     aimVector = aimDirection,
-                    damageColorIndex = DamageColorIndex.Fragile
+                    damageColorIndex = DamageColorIndex.Fragile,
+                    maxDistance = finalShotDistance
                 };
 
                 attack.damageType.damageSource = DamageSource.Secondary;
@@ -134,7 +141,10 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
                 Util.PlaySound("Play_item_use_molotov_impact_big", gameObject);
                 Util.PlayAttackSpeedSound("Play_captain_m1_hit", gameObject, 0.75f);
 
-                attack.Fire();
+                if (isAuthority)
+                {
+                    attack.Fire();
+                }
 
                 AddRecoil(4.5f, 4.5f, 0f, 0f);
                 characterMotor?.ApplyForce(-2000f * aimDirection, false, false);